requirements.txt
torch>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
ipaddress>=1.0.23 